Add demand proejction for missing countries
Some countries have no demand time series because GEGIS did not work.
We create a temporary fix in the `notebooks/0demand_gegis.ipynb`.
There we scale the timeseries of nigeria based on gdp/capita to the
respective countries. The reulting Africa.nc demand file needs to
be manually replaced in the demand data.